fix & add GameOver UI in CastleWallManager.cs
